requirements.txt
requests==2.32.3
requests-unixsocket2==0.4.2